main chart  classes move to main chart package
