+++ app/src/main/java/com/example/mementomori/MainViewModel.kt
@@ -2,7 +2,7 @@ package com.example.mementomori
 
 import androidx.lifecycle.ViewModel
 import com.example.mementomori.data.const.AgeGroups
-import com.example.mementomori.screen.UserDataViewModel
+import com.example.mementomori.screen.mainChart.UserDataViewModel
 
 class MainViewModel(
     val userDataVM: UserDataViewModel = UserDataViewModel(),
