+++ app/src/main/java/com/example/mementomori/screen/mainChart/MainChartViewModel.kt
@@ -1,10 +1,11 @@
-package com.example.mementomori.screen
+package com.example.mementomori.screen.mainChart
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 
-class UserDataViewModel: ViewModel() {
-    var userOldMonths: Int by mutableStateOf(600)
+class MainChartViewModel: ViewModel() {
+
+
 }
